+++ graphql/schema.go
@@ -17,7 +17,10 @@
 package graphql
 
 import (
+	"context"
+	"reflect"
 	"strings"
+	"sync"
 
 	"golang.org/x/xerrors"
 	"zombiezen.com/go/graphql-server/internal/gqlang"
@@ -29,6 +32,9 @@ type Schema struct {
 	mutation  *gqlType
 	types     map[string]*gqlType
 	typeOrder []string
+
+	mu      sync.RWMutex
+	goTypes map[typeKey]*typeDescriptor
 }
 
 // ParseSchema parses a GraphQL document containing type definitions.
@@ -70,6 +76,7 @@ func parseSchema(source string, internal bool) (*Schema, error) {
 		mutation:  typeMap["Mutation"],
 		types:     typeMap,
 		typeOrder: typeOrder,
+		goTypes:   make(map[typeKey]*typeDescriptor),
 	}
 	if !internal {
 		if schema.query == nil {
@@ -312,4 +319,244 @@ func (schema *Schema) Validate(query string) (*ValidatedQuery, []*ResponseError)
 		source: query,
 		doc:    doc,
 	}, nil
+}
+
+func (schema *Schema) typeDescriptor(key typeKey) *typeDescriptor {
+	if key.goType.Kind() == reflect.Interface {
+		return nil
+	}
+
+	// Fast path: descriptor already computed.
+	schema.mu.RLock()
+	desc := schema.goTypes[key]
+	schema.mu.RUnlock()
+	if desc != nil {
+		return desc
+	}
+
+	// Compute type descriptor.
+	schema.mu.Lock()
+	desc = schema.typeDescriptorLocked(key)
+	schema.mu.Unlock()
+	return desc
+}
+
+func (schema *Schema) typeDescriptorLocked(key typeKey) *typeDescriptor {
+	if key.goType.Kind() == reflect.Interface {
+		return nil
+	}
+
+	desc := schema.goTypes[key]
+	if desc != nil {
+		return desc
+	}
+	desc = &typeDescriptor{
+		fields: make(map[string]fieldDescriptor),
+	}
+	schema.goTypes[key] = desc
+
+	var structType reflect.Type
+	switch kind := key.goType.Kind(); {
+	case kind == reflect.Struct:
+		structType = key.goType
+	case kind == reflect.Ptr && key.goType.Elem().Kind() == reflect.Struct:
+		structType = key.goType.Elem()
+	}
+	for _, field := range key.gqlType.fields {
+		numMatches := 0
+		fdesc := fieldDescriptor{
+			fieldIndex:  -1,
+			methodIndex: -1,
+		}
+		lowerFieldName := toLower(field.name)
+		passSel := field.typ.selectionSetType() != nil
+		var fieldGoType reflect.Type
+		for i, n := 0, key.goType.NumMethod(); i < n; i++ {
+			meth := key.goType.Method(i)
+			if meth.PkgPath != "" {
+				// Don't consider unexported methods.
+				continue
+			}
+			if toLower(meth.Name) == lowerFieldName {
+				numMatches++
+				fdesc.methodIndex = i
+				if err := validateFieldMethodSignature(meth.Type, passSel); err != nil {
+					*desc = typeDescriptor{
+						err: xerrors.Errorf("can't use method %v.%s for field %s.%s: %w",
+							key.goType, meth.Name, key.gqlType.name, field.name, err),
+					}
+					return desc
+				}
+				fieldGoType = meth.Type.Out(0)
+			}
+		}
+		if structType != nil && len(field.args) == 0 {
+			for i, n := 0, structType.NumField(); i < n; i++ {
+				goField := structType.Field(i)
+				if goField.PkgPath != "" {
+					// Don't consider unexported fields.
+					continue
+				}
+				if toLower(goField.Name) == lowerFieldName {
+					numMatches++
+					fdesc.fieldIndex = i
+					fieldGoType = goField.Type
+					if key.goType.Kind() == reflect.Ptr {
+						// If the field is addressable, then use that as part of the type.
+						// For example, a Bar field on a *Foo may be used as a *Bar if
+						// needed. The pointer may get stripped below as part of
+						// innermostPointerType.
+						fieldGoType = reflect.PtrTo(fieldGoType)
+					}
+				}
+			}
+		}
+		if numMatches == 0 {
+			*desc = typeDescriptor{
+				err: xerrors.Errorf("no method or field found on %v for %s.%s",
+					key.goType, key.gqlType.name, field.name),
+			}
+			return desc
+		}
+		if numMatches > 1 {
+			*desc = typeDescriptor{
+				err: xerrors.Errorf("multiple methods and/or fields found on %v for %s.%s",
+					key.goType, key.gqlType.name, field.name),
+			}
+			return desc
+		}
+		// TODO(someday): Check field type for scalars.
+		if field.typ.isObject() {
+			fdesc.desc = schema.typeDescriptorLocked(typeKey{
+				goType:  innermostPointerType(fieldGoType),
+				gqlType: field.typ.obj,
+			})
+			if fdesc.desc.err != nil {
+				*desc = typeDescriptor{
+					err: xerrors.Errorf("field %s: %w", field.name, fdesc.desc.err),
+				}
+				return desc
+			}
+		}
+		desc.fields[field.name] = fdesc
+	}
+	return desc
+}
+
+// innermostPointerType returns the value's innermost pointer or interface type.
+func innermostPointerType(t reflect.Type) reflect.Type {
+	var tprev reflect.Type
+	for t.Kind() == reflect.Ptr {
+		tprev, t = t, t.Elem()
+	}
+	if tprev == nil || t.Kind() == reflect.Interface {
+		return t
+	}
+	return tprev
+}
+
+func validateFieldMethodSignature(mtype reflect.Type, passSel bool) error {
+	numIn := mtype.NumIn()
+	argIdx := 1 // skip past receiver
+	if argIdx < numIn && mtype.In(argIdx) == contextGoType {
+		argIdx++
+	}
+	if argIdx < numIn && mtype.In(argIdx) == argsGoType {
+		argIdx++
+	}
+	if passSel {
+		if argIdx < numIn && mtype.In(argIdx) == selectionSetGoType {
+			argIdx++
+		}
+	}
+	if argIdx != numIn {
+		return xerrors.New("wrong parameter signature")
+	}
+	switch mtype.NumOut() {
+	case 1:
+		if mtype.Out(0) == errorGoType {
+			return xerrors.New("return type must not be error")
+		}
+	case 2:
+		if mtype.Out(0) == errorGoType {
+			return xerrors.New("first return type must not be error")
+		}
+		if got := mtype.Out(1); got != errorGoType {
+			return xerrors.Errorf("second return type must be error (found %v)", got)
+		}
+	default:
+		return xerrors.New("wrong return signature")
+	}
+	return nil
+}
+
+// typeKey is the key to the schema's Go type cache.
+type typeKey struct {
+	goType  reflect.Type
+	gqlType *objectType
+}
+
+type typeDescriptor struct {
+	fields map[string]fieldDescriptor
+	err    error
+}
+
+type fieldDescriptor struct {
+	fieldIndex  int
+	methodIndex int
+	desc        *typeDescriptor
+}
+
+func (fdesc fieldDescriptor) read(ctx context.Context, recv reflect.Value, args map[string]Value, sel *SelectionSet) (reflect.Value, error) {
+	if fdesc.fieldIndex != -1 {
+		recv = unwrapPointer(recv)
+		if !recv.IsValid() {
+			return reflect.Value{}, xerrors.New("nil pointer")
+		}
+		return recv.Field(fdesc.fieldIndex), nil
+	}
+	method := recv.Method(fdesc.methodIndex)
+	mtype := method.Type()
+	numIn := mtype.NumIn()
+	var callArgs []reflect.Value
+	if len(callArgs) < numIn && mtype.In(len(callArgs)) == contextGoType {
+		callArgs = append(callArgs, reflect.ValueOf(ctx))
+	}
+	if len(callArgs) < numIn && mtype.In(len(callArgs)) == argsGoType {
+		callArgs = append(callArgs, reflect.ValueOf(args))
+	}
+	if len(callArgs) < numIn && mtype.In(len(callArgs)) == selectionSetGoType {
+		callArgs = append(callArgs, reflect.ValueOf(sel))
+	}
+	if len(callArgs) != numIn {
+		panic("unexpected method parameters; bug in validateFieldMethodSignature?")
+	}
+	switch mtype.NumOut() {
+	case 1:
+		out := method.Call(callArgs)
+		return out[0], nil
+	case 2:
+		out := method.Call(callArgs)
+		if !out[1].IsNil() {
+			// Intentionally making the returned error opaque to avoid interference in
+			// toResponseError.
+			err := out[1].Interface().(error)
+			return reflect.Value{}, xerrors.Errorf("server error: %v", err)
+		}
+		return out[0], nil
+	default:
+		panic("unexpected method return signature; bug in validateFieldMethodSignature?")
+	}
+}
+
+func toLower(s string) string {
+	sb := new(strings.Builder)
+	for i := 0; i < len(s); i++ {
+		if 'A' <= s[i] && s[i] <= 'Z' {
+			sb.WriteByte(s[i] - 'A' + 'a')
+		} else {
+			sb.WriteByte(s[i])
+		}
+	}
+	return sb.String()
 }