+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -  Marshaling of Go types into GraphQL output types
 -  Context propagation
 -  Precise error reporting
+-  Verification of Go types as result types
 
 ### Not Implemented
 
@@ -33,6 +34,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -  Concurrent field resolution
 -  Validation for custom scalar types
 -  Unmarshaling of arguments into Go types
--  Verification of Go types as result types
 -  Metrics and trace spans
 -  Explicit `schema` blocks